update test errors to be easier to work with so that new cases don't shift line numbers around
