+++ tests/root.go
@@ -5,8 +5,10 @@ import (
 	"fmt"
 
 	"github.com/Southclaws/fault"
-	"github.com/Southclaws/fault/fmsg"
 )
+
+// the root cause of errors for testing. All line numbers here should remain the
+// same so the tests don't have to be edited constantly when new cases are added
 
 var (
 	errSentinelStdlib = errors.New("stdlib sentinel error")
@@ -27,32 +29,5 @@ func rootCause(kind int) error {
 	} else if kind == 5 {
 		return fmt.Errorf("errorf wrapped: %w", errSentinelStdlib)
 	}
-	return nil
-}
-
-func errorCallerDeep(kind int) error {
-	err := rootCause(kind)
-	if err != nil {
-		return fault.Wrap(err)
-	}
-
-	return nil
-}
-
-func errorCallerMid(kind int) error {
-	err := errorCallerDeep(kind)
-	if err != nil {
-		return fault.Wrap(err, fmsg.With("failed to call function"))
-	}
-
-	return nil
-}
-
-func errorCaller(kind int) error {
-	err := errorCallerMid(kind)
-	if err != nil {
-		return fault.Wrap(err)
-	}
-
 	return nil
 }
